updated petserviceid property in domain models

+++ DatamartManagementService/DatamartManagementService.Domain/Models/EmployeePayrollDetail.cs
@@ -12,7 +12,7 @@ namespace DatamartManagementService.Domain.Models
         
         public string LastName { get; set; }
         
-        public long PetServiceId { get; set; }
+        public short PetServiceId { get; set; }
         
         public string PetServiceName { get; set; }
